Description of files
Their are four files, in which data description contains the description of rows and column of the data. Train and test files are in CSV which contains all the data. The actual work has been done in the Task HD file that contains whole data science project from data acquisition till predictive modeling.